refactor: 유효성 검사 로직 utils 폴더로 이동
