--- src/constants/errorMessages.js
@@ -1,7 +1,0 @@
-// errorMessages.js
-export const ERROR_MESSAGES = {
-  INVALID_NUMBER: '숫자를 전달해야 합니다.',
-  INVALID_RANGE: '첫 번째 인자가 두 번째 인자보다 작거나 같아야 합니다.',
-  INVALID_FUNCTION: '함수를 전달해야 합니다.',
-  INVALID_TIMES_RANGE: '횟수는 1 이상이어야 합니다.',
-};
